+++ day2/section03-paxos/src/main/java/com/distribpatterns/paxos/PrepareResponse.java
@@ -13,7 +13,7 @@ package com.distribpatterns.paxos;
  */
 public record PrepareResponse(
     boolean promised,
-    Integer acceptedGeneration,
+    Long acceptedGeneration,
     Operation acceptedValue
 ) {
 }
