+++ day2/section03-paxos/src/main/java/com/distribpatterns/paxos/ProposeRequest.java
@@ -4,6 +4,6 @@ package com.distribpatterns.paxos;
  * Phase 2a message: Proposer asks acceptors to accept this specific value
  * for the generation.
  */
-public record ProposeRequest(int generation, Operation value) {
+public record ProposeRequest(long generation, Operation value) {
 }
